+++ src/main/java/com/server/EZY/exception/customError/exception/CustomNotFoundException.java
@@ -1,4 +1,4 @@
-package com.server.EZY.exception.custom.exception;
+package com.server.EZY.exception.customError.exception;
 
 public class CustomNotFoundException extends RuntimeException{
     public CustomNotFoundException(String msg, Throwable t){
